perf(homepage): server-render CuratedSystems + defer FloatingBar island
- CuratedSystems was a client component fetching /api/curated-systems in a
  useEffect → skeleton → content swap, the likely source of the field CLS
  (0.11) that doesn't reproduce on localhost. Extract the query into a cached
  lib (getCuratedSystems, unstable_cache 1h, shared with the API route) and
  render the section server-side; only the cans/iems tab toggle stays client
  (CuratedSystemsClient). The fully-laid-out grid now ships in initial HTML —
  no fetch, no skeleton, no swap.
- FloatingBar: load client-only via next/dynamic ssr:false (FloatingBarLazy).
  It's a scroll-triggered fixed bar, invisible at first paint and SEO-irrelevant,
  so it no longer ships in SSR or hydrates on the critical path. Fixed-position
  → no layout shift on mount.

Verified: build passes; SSR HTML contains "Popular systems" + real curated card
hrefs (no skeleton); /api/curated-systems still returns data; Lighthouse desktop
Perf 100, CLS 0, LCP 0.74s.
